move tests back to unit tests
